rework config lt tests
